V2.3 release
Switched a resistor and capacitor to B.cu to ensure they don't get ripped off when removing the picoblade connector. depending on further testing to see if I can get the LEDout working on the SEC chip, might do the same with R1 and remove D1 and D2 and the associated resistor. 
Cleaned up routing a bit due to that change.