removed no longer needed features (LEDs, USB)
After testing it turns out the USB port on the SEC1210 UART version does not seem to be enabled. Therefore there's no way to program the chip to enable the LED output.
Removed unneeded D1, D2, R6, and re-positioned R1 and C1 thanks to the extra space.
updated schematic with DNP tags for D1, D2, R6.